Fixed laser spawning amount error
This was not a glitch in the game. This was actually a user error in the Laser Tatsu pattern (which was the only one used for testing because of its playability). All of the laser nodes were placed down manually, instead of one that points to where it should shoot. This was the cause of the "glitch"; it was just a user error.